server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.23
argon2-cffi==23.1.0
python-dotenv==1.0.0
marshmallow==3.20.1